+++ python/mozbuild/mozpack/files.py
@@ -4,6 +4,7 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+import codecs
 import errno
 import inspect
 import os
@@ -18,7 +19,6 @@ from itertools import chain
 from mozbuild.preprocessor import Preprocessor
 from mozbuild.util import (
     FileAvoidWrite,
-    ensure_bytes,
     ensure_unicode,
 )
 from mozpack.executables import (
@@ -76,6 +76,14 @@ else:
             raise TypeError('mismatched path types!')
 
 
+# Helper function; ensures we always open files with the correct encoding when
+# opening them in text mode.
+def _open(path, mode='r'):
+    if six.PY3 and 'b' not in mode:
+        return open(path, mode, encoding='utf-8')
+    return open(path, mode)
+
+
 class Dest(object):
     '''
     Helper interface for BaseFile.copy. The interface works as follows:
@@ -88,8 +96,9 @@ class Dest(object):
     '''
 
     def __init__(self, path):
-        self.path = ensure_unicode(path)
+        self.file = None
         self.mode = None
+        self.path = ensure_unicode(path)
 
     @property
     def name(self):
@@ -97,15 +106,16 @@ class Dest(object):
 
     def read(self, length=-1):
         if self.mode != 'r':
-            self.file = open(self.path, 'rb')
+            self.file = _open(self.path, mode='rb')
             self.mode = 'r'
         return self.file.read(length)
 
     def write(self, data):
         if self.mode != 'w':
-            self.file = open(self.path, 'wb')
+            self.file = _open(self.path, mode='wb')
             self.mode = 'w'
-        return self.file.write(data)
+        to_write = six.ensure_binary(data)
+        return self.file.write(to_write)
 
     def exists(self):
         return os.path.exists(self.path)
@@ -114,6 +124,7 @@ class Dest(object):
         if self.mode:
             self.mode = None
             self.file.close()
+            self.file = None
 
 
 class BaseFile(object):
@@ -200,6 +211,13 @@ class BaseFile(object):
 
         if can_skip_content_check:
             if getattr(self, 'path', None) and getattr(dest, 'path', None):
+                # The destination directory must exist, or CopyFile will fail.
+                destdir = os.path.dirname(dest.path)
+                try:
+                    os.makedirs(destdir)
+                except OSError as e:
+                    if e.errno != errno.EEXIST:
+                        raise
                 _copyfile(self.path, dest.path)
                 shutil.copystat(self.path, dest.path)
             else:
@@ -210,17 +228,18 @@ class BaseFile(object):
             return True
 
         src = self.open()
-        copy_content = b''
+        accumulated_src_content = []
         while True:
             dest_content = dest.read(32768)
             src_content = src.read(32768)
-            copy_content += src_content
+            accumulated_src_content.append(src_content)
             if len(dest_content) == len(src_content) == 0:
                 break
             # If the read content differs between origin and destination,
             # write what was read up to now, and copy the remainder.
-            if dest_content != src_content:
-                dest.write(copy_content)
+            if (six.ensure_binary(dest_content) !=
+                six.ensure_binary(src_content)):
+                dest.write(b''.join(accumulated_src_content))
                 shutil.copyfileobj(src, dest)
                 break
         if hasattr(self, 'path') and hasattr(dest, 'path'):
@@ -356,7 +375,13 @@ class AbsoluteSymlinkFile(File):
 
         # Handle the simple case where symlinks are definitely not supported by
         # falling back to file copy.
-        if not hasattr(os, 'symlink'):
+        # Python 3 supports symlinks on Windows, but for some reason, some JS
+        # files fail to be loaded with permission errors only when they are
+        # symlinks as of May 2020 update. Exclude them for now until we figure
+        # out the cause.
+        js_on_windows = platform.system() == 'Windows' and \
+            self.path.endswith((".js", ".jsm"))
+        if not hasattr(os, 'symlink') or js_on_windows:
             return File.copy(self, dest, skip_if_older=skip_if_older)
 
         # Always verify the symlink target path exists.
@@ -536,8 +561,8 @@ class PreprocessedFile(BaseFile):
         pp = Preprocessor(defines=self.defines, marker=self.marker)
         pp.setSilenceDirectiveWarnings(self.silence_missing_directive_warnings)
 
-        with open(self.path, 'rU') as input:
-            with open(os.devnull, 'w') as output:
+        with _open(self.path, 'rU') as input:
+            with _open(os.devnull, 'w') as output:
                 pp.processFile(input=input, output=output)
 
         # This always yields at least self.path.
@@ -558,7 +583,8 @@ class PreprocessedFile(BaseFile):
         # destination is not a symlink, we leave it alone, since we're going to
         # overwrite its contents anyway.
         # If symlinks aren't supported at all, we can skip this step.
-        if hasattr(os, 'symlink'):
+        # See comment in AbsoluteSymlinkFile about Windows.
+        if hasattr(os, 'symlink') and platform.system() != 'Windows':
             if os.path.islink(dest.path):
                 os.remove(dest.path)
 
@@ -568,7 +594,7 @@ class PreprocessedFile(BaseFile):
         # dependencies from that file to our list.
         if self.depfile and os.path.exists(self.depfile):
             target = mozpath.normpath(dest.name)
-            with open(self.depfile, 'rb') as fileobj:
+            with _open(self.depfile, 'rt') as fileobj:
                 for rule in makeutil.read_dep_makefile(fileobj):
                     if target in rule.targets():
                         pp_deps.update(rule.dependencies())
@@ -592,7 +618,7 @@ class PreprocessedFile(BaseFile):
         pp = Preprocessor(defines=self.defines, marker=self.marker)
         pp.setSilenceDirectiveWarnings(self.silence_missing_directive_warnings)
 
-        with open(self.path, 'rU') as input:
+        with _open(self.path, 'rU') as input:
             pp.processFile(input=input, output=dest, depfile=deps_out)
 
         dest.close()
@@ -614,7 +640,7 @@ class GeneratedFile(BaseFile):
     def content(self):
         if inspect.isfunction(self._content):
             self._content = self._content()
-        return self._content
+        return six.ensure_binary(self._content)
 
     @content.setter
     def content(self, content):
@@ -659,11 +685,11 @@ class ExtractedTarFile(GeneratedFile):
         assert isinstance(info, TarInfo)
         assert isinstance(tar, TarFile)
         GeneratedFile.__init__(self, tar.extractfile(info).read())
-        self._mode = self.normalize_mode(info.mode)
+        self._unix_mode = self.normalize_mode(info.mode)
 
     @property
     def mode(self):
-        return self._mode
+        return self._unix_mode
 
     def read(self):
         return self.content
@@ -718,12 +744,10 @@ class ManifestFile(BaseFile):
         Return a file-like object allowing to read() the serialized content of
         the manifest.
         '''
-        return BytesIO(
-            ensure_bytes(
-                ''.join(
-                    '%s\n' % e.rebase(self._base)
-                    for e in chain(self._entries, self._interfaces)
-                )))
+        content = ''.join(
+            '%s\n' % e.rebase(self._base)
+            for e in chain(self._entries, self._interfaces))
+        return BytesIO(six.ensure_binary(content))
 
     def __iter__(self):
         '''
@@ -753,8 +777,11 @@ class MinifiedProperties(BaseFile):
         Return a file-like object allowing to read() the minified content of
         the properties file.
         '''
-        return BytesIO(b''.join(l for l in self._file.open().readlines()
-                                if not l.startswith(b'#')))
+        content = ''.join(
+            l for l in [
+                six.ensure_text(s) for s in self._file.open().readlines()
+            ] if not l.startswith('#'))
+        return BytesIO(six.ensure_binary(content))
 
 
 class MinifiedJavaScript(BaseFile):
@@ -768,18 +795,20 @@ class MinifiedJavaScript(BaseFile):
         self._verify_command = verify_command
 
     def open(self):
-        output = BytesIO()
-        minify = JavascriptMinify(self._file.open(), output, quote_chars="'\"`")
+        output = six.StringIO()
+        minify = JavascriptMinify(codecs.getreader('utf-8')(self._file.open()),
+                                  output, quote_chars="'\"`")
         minify.minify()
         output.seek(0)
+        output_source = six.ensure_binary(output.getvalue())
+        output = BytesIO(output_source)
 
         if not self._verify_command:
             return output
 
         input_source = self._file.open().read()
-        output_source = output.getvalue()
 
-        with NamedTemporaryFile() as fh1, NamedTemporaryFile() as fh2:
+        with NamedTemporaryFile('wb+') as fh1, NamedTemporaryFile('wb+') as fh2:
             fh1.write(input_source)
             fh2.write(output_source)
             fh1.flush()
@@ -788,7 +817,8 @@ class MinifiedJavaScript(BaseFile):
             try:
                 args = list(self._verify_command)
                 args.extend([fh1.name, fh2.name])
-                subprocess.check_output(args, stderr=subprocess.STDOUT)
+                subprocess.check_output(args, stderr=subprocess.STDOUT,
+                                        universal_newlines=True)
             except subprocess.CalledProcessError as e:
                 errors.warn('JS minification verification failed for %s:' %
                             (getattr(self._file, 'path', '<unknown>')))
@@ -1123,7 +1153,11 @@ class MercurialFile(BaseFile):
     """File class for holding data from Mercurial."""
 
     def __init__(self, client, rev, path):
-        self._content = client.cat([path], rev=rev)
+        self._content = client.cat([six.ensure_binary(path)],
+                                   rev=six.ensure_binary(rev))
+
+    def open(self):
+        return BytesIO(six.ensure_binary(self._content))
 
     def read(self):
         return self._content
@@ -1162,16 +1196,18 @@ class MercurialRevisionFinder(BaseFinder):
             self._client = hglib.open(path=repo, encoding=b'utf-8')
         finally:
             os.chdir(oldcwd)
-        self._rev = rev if rev is not None else b'.'
+        self._rev = rev if rev is not None else '.'
         self._files = OrderedDict()
 
         # Immediately populate the list of files in the repo since nearly every
         # operation requires this list.
-        out = self._client.rawcommand([b'files', b'--rev', str(self._rev)])
+        out = self._client.rawcommand([
+            b'files', b'--rev', six.ensure_binary(self._rev),
+        ])
         for relpath in out.splitlines():
             # Mercurial may use \ as path separator on Windows. So use
             # normpath().
-            self._files[mozpath.normpath(relpath)] = None
+            self._files[six.ensure_text(mozpath.normpath(relpath))] = None
 
     def _find(self, pattern):
         if self._recognize_repo_paths:
